+++ requirements.txt
@@ -1,0 +1,2 @@
+pyyaml==5.3
+clikit==0.4.1
