+++ tox.ini
@@ -1,0 +1,23 @@
+[tox]
+envlist =
+       {py36}-django22
+       {py36,py37,py38}-django30,
+       {py36,py37,py38}-djangomaster,
+
+[travis:env]
+DJANGO =
+    2.2: django22
+    3.0: django30
+    master: djangomaster
+
+[testenv]
+commands = python manage.py test tivol/tests
+envdir = {toxworkdir}/venvs/{envname}
+setenv =
+       PYTHONDONTWRITEBYTECODE=1
+       PYTHONWARNINGS=once
+deps =
+        django22: Django>=2.2,<3.0
+        django30: Django>=3.0,<3.1
+        djangomaster: https://github.com/django/django/archive/master.tar.gz
+        -rrequirements.txt
